Consolidate RPC/REST duplicated tests into client tests
