+++ remote/client_windows_test.go
