+++ remote/rest_service_test.go
@@ -5,18 +5,15 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
-	"net"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"path/filepath"
-	"runtime"
 	"strings"
 	"syscall"
 	"testing"
 	"time"
 
-	"github.com/mongodb/grip"
 	"github.com/mongodb/jasper"
 	"github.com/mongodb/jasper/mock"
 	"github.com/mongodb/jasper/options"
@@ -37,9 +34,9 @@ func TestRestService(t *testing.T) {
 	httpClient := testutil.GetHTTPClient()
 	defer testutil.PutHTTPClient(httpClient)
 
-	tempDir, err := ioutil.TempDir("./", "jasper-rest-service-test")
+	tempDir, err := ioutil.TempDir(buildDir(t), filepath.Base(t.Name()))
 	require.NoError(t, err)
-	defer func() { require.NoError(t, os.RemoveAll(tempDir)) }()
+	defer func() { assert.NoError(t, os.RemoveAll(tempDir)) }()
 
 	for name, test := range map[string]func(context.Context, *testing.T, *Service, *restClient){
 		"VerifyFixtures": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
@@ -56,30 +53,7 @@ func TestRestService(t *testing.T) {
 			assert.Error(t, err)
 			assert.Error(t, handleError(&http.Response{Body: &neverJSON{}, StatusCode: http.StatusTeapot}))
 		},
-		"EmptyCreateOpts": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			proc, err := client.CreateProcess(ctx, &options.Create{})
-			assert.Error(t, err)
-			assert.Nil(t, proc)
-		},
-		"WithOnlyTimeoutValue": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			proc, err := client.CreateProcess(ctx, &options.Create{Args: []string{"ls"}, TimeoutSecs: 300})
-			assert.NoError(t, err)
-			assert.NotNil(t, proc)
-		},
-		"ListErrorsWithInvalidFilter": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			list, err := client.List(ctx, "foo")
-			assert.Error(t, err)
-			assert.Nil(t, list)
-		},
-		"RegisterAlwaysErrors": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			proc, err := client.CreateProcess(ctx, &options.Create{Args: []string{"ls"}})
-			assert.NotNil(t, proc)
-			require.NoError(t, err)
-
-			assert.Error(t, client.Register(ctx, nil))
-			assert.Error(t, client.Register(ctx, proc))
-		},
-		"ClientMethodsErrorWithBadUrl": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
+		"ClientMethodsErrorWithBadURL": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			client.prefix = strings.Replace(client.prefix, "http://", "://", 1)
 
 			_, err := client.List(ctx, options.All)
@@ -173,7 +147,7 @@ func TestRestService(t *testing.T) {
 			assert.Error(t, err)
 			assert.Contains(t, err.Error(), "problem making request")
 		},
-		"ProcessMethodsWithBadUrl": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
+		"ProcessMethodsWithBadURL": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			client.prefix = strings.Replace(client.prefix, "http://", "://", 1)
 
 			proc := &restProcess{
@@ -250,7 +224,6 @@ func TestRestService(t *testing.T) {
 			err := proc.Signal(ctx, syscall.SIGTERM)
 			assert.Error(t, err)
 			assert.Contains(t, err.Error(), "no process")
-
 		},
 		"CreateProcessEndpointErrorsWithMalformedData": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			body, err := makeBody(map[string]int{"tags": 42})
@@ -261,82 +234,6 @@ func TestRestService(t *testing.T) {
 			rw := httptest.NewRecorder()
 			srv.createProcess(rw, req)
 			assert.Equal(t, http.StatusBadRequest, rw.Code)
-		},
-		"StandardInput": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			for subTestName, subTestCase := range map[string]func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte){
-				"ReaderIsIgnored": func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte) {
-					opts.StandardInput = bytes.NewBuffer(stdin)
-
-					proc, err := client.CreateProcess(ctx, opts)
-					require.NoError(t, err)
-
-					_, err = proc.Wait(ctx)
-					require.NoError(t, err)
-
-					logs, err := client.GetLogStream(ctx, proc.ID(), 1)
-					require.NoError(t, err)
-					assert.Empty(t, logs.Logs)
-				},
-				"BytesSetsStandardInput": func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte) {
-					opts.StandardInputBytes = stdin
-
-					proc, err := client.CreateProcess(ctx, opts)
-					require.NoError(t, err)
-
-					_, err = proc.Wait(ctx)
-					require.NoError(t, err)
-
-					logs, err := client.GetLogStream(ctx, proc.ID(), 1)
-					require.NoError(t, err)
-
-					require.Len(t, logs.Logs, 1)
-					assert.Equal(t, expectedOutput, strings.TrimSpace(logs.Logs[0]))
-				},
-				"BytesCopiedByRespawnedProcess": func(ctx context.Context, t *testing.T, opts *options.Create, expectedOutput string, stdin []byte) {
-					opts.StandardInputBytes = stdin
-
-					proc, err := client.CreateProcess(ctx, opts)
-					require.NoError(t, err)
-
-					_, err = proc.Wait(ctx)
-					require.NoError(t, err)
-
-					logs, err := client.GetLogStream(ctx, proc.ID(), 1)
-					require.NoError(t, err)
-
-					require.Len(t, logs.Logs, 1)
-					assert.Equal(t, expectedOutput, strings.TrimSpace(logs.Logs[0]))
-
-					newProc, err := proc.Respawn(ctx)
-					require.NoError(t, err)
-
-					_, err = newProc.Wait(ctx)
-					require.NoError(t, err)
-
-					logs, err = client.GetLogStream(ctx, newProc.ID(), 1)
-					require.NoError(t, err)
-
-					require.Len(t, logs.Logs, 1)
-					assert.Equal(t, expectedOutput, strings.TrimSpace(logs.Logs[0]))
-				},
-			} {
-				t.Run(subTestName, func(t *testing.T) {
-					opts := &options.Create{
-						Args: []string{"bash", "-s"},
-						Output: options.Output{
-							Loggers: []options.Logger{jasper.NewInMemoryLogger(1)},
-						},
-					}
-					expectedOutput := "foobar"
-					stdin := []byte("echo " + expectedOutput)
-					subTestCase(ctx, t, opts, expectedOutput, stdin)
-				})
-			}
-		},
-		"InvalidFilterReturnsError": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			procs, err := client.List(ctx, options.Filter("foo"))
-			assert.Error(t, err)
-			assert.Nil(t, procs)
 		},
 		"WaitForProcessThatDoesNotExistShouldError": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			proc := &restProcess{
@@ -439,7 +336,7 @@ func TestRestService(t *testing.T) {
 			assert.NoError(t, err)
 
 		},
-		"SignalFailsToParsePid": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
+		"SignalFailsToParsePID": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			req, err := http.NewRequest(http.MethodPatch, client.getURL("/process/%s/signal/f", "foo"), nil)
 			require.NoError(t, err)
 			req = req.WithContext(ctx)
@@ -449,145 +346,6 @@ func TestRestService(t *testing.T) {
 			defer resp.Body.Close()
 			assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
 			assert.Contains(t, handleError(resp).Error(), "problem converting signal 'f'")
-		},
-		"DownloadFileCreatesResource": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			file, err := ioutil.TempFile(tempDir, "out.txt")
-			require.NoError(t, err)
-			defer func() {
-				assert.NoError(t, file.Close())
-				assert.NoError(t, os.RemoveAll(file.Name()))
-			}()
-			absPath, err := filepath.Abs(file.Name())
-			require.NoError(t, err)
-
-			assert.NoError(t, client.DownloadFile(ctx, options.Download{URL: "https://example.com", Path: absPath}))
-
-			info, err := os.Stat(file.Name())
-			assert.NoError(t, err)
-			assert.NotEqual(t, 0, info.Size())
-		},
-		"DownloadFileCreatesResourceAndExtracts": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			downloadDir, err := ioutil.TempDir(tempDir, "rest_test")
-			require.NoError(t, err)
-			defer func() {
-				assert.NoError(t, os.RemoveAll(downloadDir))
-			}()
-
-			fileServerDir, err := ioutil.TempDir(tempDir, "rest_test_server")
-			require.NoError(t, err)
-			defer func() {
-				assert.NoError(t, os.RemoveAll(fileServerDir))
-			}()
-
-			fileName := "foo.zip"
-			fileContents := "foo"
-			require.NoError(t, testutil.AddFileToDirectory(fileServerDir, fileName, fileContents))
-
-			absDownloadDir, err := filepath.Abs(downloadDir)
-			require.NoError(t, err)
-			destFilePath := filepath.Join(absDownloadDir, fileName)
-			destExtractDir := filepath.Join(absDownloadDir, "extracted")
-
-			port := testutil.GetPortNumber()
-			fileServerAddr := fmt.Sprintf("localhost:%d", port)
-			fileServer := &http.Server{Addr: fileServerAddr, Handler: http.FileServer(http.Dir(fileServerDir))}
-			defer func() {
-				assert.NoError(t, fileServer.Close())
-			}()
-			listener, err := net.Listen("tcp", fileServerAddr)
-			require.NoError(t, err)
-			go func() {
-				grip.Info(fileServer.Serve(listener))
-			}()
-
-			baseURL := fmt.Sprintf("http://%s", fileServerAddr)
-			require.NoError(t, testutil.WaitForRESTService(ctx, baseURL))
-
-			opts := options.Download{
-				URL:  fmt.Sprintf("%s/%s", baseURL, fileName),
-				Path: destFilePath,
-				ArchiveOpts: options.Archive{
-					ShouldExtract: true,
-					Format:        options.ArchiveZip,
-					TargetPath:    destExtractDir,
-				},
-			}
-			require.NoError(t, client.DownloadFile(ctx, opts))
-
-			fileInfo, err := os.Stat(destFilePath)
-			require.NoError(t, err)
-			assert.NotZero(t, fileInfo.Size())
-
-			dirContents, err := ioutil.ReadDir(destExtractDir)
-			require.NoError(t, err)
-
-			assert.NotZero(t, len(dirContents))
-		},
-		"DownloadFileFailsExtractionWithInvalidArchiveFormat": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			fileName := filepath.Join(tempDir, "out.txt")
-			_, err := os.Stat(fileName)
-			require.True(t, os.IsNotExist(err))
-
-			opts := options.Download{
-				URL:  "https://example.com",
-				Path: fileName,
-				ArchiveOpts: options.Archive{
-					ShouldExtract: true,
-					Format:        options.ArchiveFormat("foo"),
-				},
-			}
-			assert.Error(t, client.DownloadFile(ctx, opts))
-
-			_, err = os.Stat(fileName)
-			assert.True(t, os.IsNotExist(err))
-		},
-		"DownloadFileFailsForUnarchivedFile": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			file, err := ioutil.TempFile(tempDir, "out.txt")
-			require.NoError(t, err)
-			defer func() {
-				assert.NoError(t, file.Close())
-				assert.NoError(t, os.RemoveAll(file.Name()))
-			}()
-			extractDir, err := ioutil.TempDir(tempDir, "out")
-			require.NoError(t, err)
-			defer func() {
-				assert.NoError(t, os.RemoveAll(extractDir))
-			}()
-
-			opts := options.Download{
-				URL:  "https://example.com",
-				Path: file.Name(),
-				ArchiveOpts: options.Archive{
-					ShouldExtract: true,
-					Format:        options.ArchiveAuto,
-				},
-			}
-			assert.Error(t, client.DownloadFile(ctx, opts))
-
-			dirContents, err := ioutil.ReadDir(extractDir)
-			require.NoError(t, err)
-			assert.Zero(t, len(dirContents))
-		},
-		"DownloadFileFailsWithInvalidURL": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			err := client.DownloadFile(ctx, options.Download{URL: "", Path: ""})
-			assert.Error(t, err)
-		},
-		"DownloadFileFailsForNonexistentURL": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			file, err := ioutil.TempFile(tempDir, "out.txt")
-			require.NoError(t, err)
-			defer func() {
-				assert.NoError(t, file.Close())
-				assert.NoError(t, os.RemoveAll(file.Name()))
-			}()
-			assert.Error(t, client.DownloadFile(ctx, options.Download{URL: "https://example.com/foo", Path: file.Name()}))
-		},
-		"DownloadFileFailsForInsufficientPermissions": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			if os.Geteuid() == 0 {
-				t.Skip("cannot test download permissions as root")
-			} else if runtime.GOOS == "windows" {
-				t.Skip("cannot test download permissions on windows")
-			}
-			assert.Error(t, client.DownloadFile(ctx, options.Download{URL: "https://example.com", Path: "/foo/bar"}))
 		},
 		"ServiceDownloadFileFailsWithInvalidOptions": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			body, err := makeBody(struct {
@@ -600,70 +358,6 @@ func TestRestService(t *testing.T) {
 			rw := httptest.NewRecorder()
 			srv.downloadFile(rw, req)
 			assert.Equal(t, http.StatusBadRequest, rw.Code)
-		},
-		"ServiceDownloadFileFailsWithInvalidURL": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			fileName := filepath.Join(tempDir, "out.txt")
-			absPath, err := filepath.Abs(fileName)
-			require.NoError(t, err)
-
-			body, err := makeBody(options.Download{
-				URL:  "://example.com",
-				Path: absPath,
-			})
-			require.NoError(t, err)
-
-			req, err := http.NewRequest(http.MethodPost, client.getURL("/download"), body)
-			require.NoError(t, err)
-			rw := httptest.NewRecorder()
-
-			srv.downloadFile(rw, req)
-			assert.Equal(t, http.StatusInternalServerError, rw.Code)
-		},
-		"WithInMemoryLogger": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			output := "foo"
-			opts := &options.Create{
-				Args: []string{"echo", output},
-				Output: options.Output{
-					Loggers: []options.Logger{
-						{
-							Type:    options.LogInMemory,
-							Options: options.Log{InMemoryCap: 100, Format: options.LogFormatPlain},
-						},
-					},
-				},
-			}
-
-			for testName, testCase := range map[string]func(ctx context.Context, t *testing.T, proc jasper.Process){
-				"GetLogStreamFailsForInvalidCount": func(ctx context.Context, t *testing.T, proc jasper.Process) {
-					stream, err := client.GetLogStream(ctx, proc.ID(), -1)
-					assert.Error(t, err)
-					assert.Zero(t, stream)
-				},
-				"GetLogStreamReturnsOutputOnSuccess": func(ctx context.Context, t *testing.T, proc jasper.Process) {
-					logs := []string{}
-					for stream, err := client.GetLogStream(ctx, proc.ID(), 1); !stream.Done; stream, err = client.GetLogStream(ctx, proc.ID(), 1) {
-						require.NoError(t, err)
-						require.NotEmpty(t, stream.Logs)
-						logs = append(logs, stream.Logs...)
-					}
-					assert.Contains(t, logs, output)
-				},
-			} {
-				t.Run(testName, func(t *testing.T) {
-					proc, err := client.CreateProcess(ctx, opts)
-					require.NoError(t, err)
-					require.NotNil(t, proc)
-
-					_, err = proc.Wait(ctx)
-					require.NoError(t, err)
-					testCase(ctx, t, proc)
-				})
-			}
-		},
-		"GetLogStreamFromNonexistentProcessFails": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			stream, err := client.GetLogStream(ctx, "foo", 1)
-			assert.Error(t, err)
-			assert.Zero(t, stream)
 		},
 		"GetLogStreamFailsWithoutInMemoryLogger": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			opts := &options.Create{Args: []string{"echo", "foo"}}
@@ -746,23 +440,6 @@ func TestRestService(t *testing.T) {
 
 			err = client.DownloadMongoDB(ctx, opts)
 			assert.NoError(t, err)
-		},
-		"GetBuildloggerURLsFailsWithNonexistentProcess": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			urls, err := client.GetBuildloggerURLs(ctx, "foo")
-			assert.Error(t, err)
-			assert.Nil(t, urls)
-		},
-		"GetBuildloggerURLsFailsWithoutBuildlogger": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			opts := &options.Create{Args: []string{"echo", "foo"}}
-			opts.Output.Loggers = []options.Logger{{Type: options.LogDefault, Options: options.Log{Format: options.LogFormatPlain}}}
-
-			proc, err := client.CreateProcess(ctx, opts)
-			assert.NoError(t, err)
-			assert.NotNil(t, proc)
-
-			urls, err := client.GetBuildloggerURLs(ctx, proc.ID())
-			assert.Error(t, err)
-			assert.Nil(t, urls)
 		},
 		"CreateWithMultipleLoggers": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			file, err := ioutil.TempFile(tempDir, "out.txt")
@@ -856,26 +533,7 @@ func TestRestService(t *testing.T) {
 
 			assert.Equal(t, buf, content)
 		},
-		"WriteFileFailsWithInvalidPath": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			opts := options.WriteFile{Content: []byte("foo")}
-			assert.Error(t, client.WriteFile(ctx, opts))
-		},
-		"WriteFileSucceedsWithNoContent": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			path := filepath.Join(tempDir, "write_file")
-			require.NoError(t, os.RemoveAll(path))
-			defer func() {
-				assert.NoError(t, os.RemoveAll(path))
-			}()
-
-			opts := options.WriteFile{Path: path}
-			require.NoError(t, client.WriteFile(ctx, opts))
-
-			stat, err := os.Stat(path)
-			require.NoError(t, err)
-
-			assert.Zero(t, stat.Size())
-		},
-		"ServiceRegisterSignalTriggerIDChecksForExistingProcess": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
+		"RegisterSignalTriggerIDChecksForExistingProcess": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
 			req, err := http.NewRequest(http.MethodPatch, client.getURL("/process/%s/trigger/signal/%s", "foo", jasper.CleanTerminationSignalTrigger), nil)
 			require.NoError(t, err)
 
@@ -886,25 +544,6 @@ func TestRestService(t *testing.T) {
 			assert.Equal(t, http.StatusBadRequest, resp.StatusCode)
 			assert.Contains(t, handleError(resp).Error(), "no process 'foo' found")
 		},
-		"ServiceRegisterSignalTriggerIDChecksForInvalidTriggerID": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			proc, err := client.CreateProcess(ctx, testutil.YesCreateOpts(0))
-			require.NoError(t, err)
-			assert.True(t, proc.Running(ctx))
-
-			assert.Error(t, proc.RegisterSignalTriggerID(ctx, jasper.SignalTriggerID("foo")))
-
-			assert.NoError(t, proc.Signal(ctx, syscall.SIGTERM))
-		},
-		"ServiceRegisterSignalTriggerIDPassesWithValidArgs": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {
-			proc, err := client.CreateProcess(ctx, testutil.YesCreateOpts(0))
-			require.NoError(t, err)
-			assert.True(t, proc.Running(ctx))
-
-			assert.NoError(t, proc.RegisterSignalTriggerID(ctx, jasper.CleanTerminationSignalTrigger))
-
-			assert.NoError(t, proc.Signal(ctx, syscall.SIGTERM))
-		},
-		// "": func(ctx context.Context, t *testing.T, srv *Service, client *restClient) {},
 	} {
 		t.Run(name, func(t *testing.T) {
 			ctx, cancel := context.WithTimeout(context.Background(), testutil.LongTestTimeout)
